Update get_token_ids.py to better support opensource LLMs using env var: MOCK_OPENAI_LLM
Each of the previous get_token_id functions would throw a KeyError if a model was set in the config and not found in the "encodings" and "max_length" dictionaries.  These updated functions now allow the model in the config to be set to an opensource LLM as long as the "MOCK_OPENAI_LLM" environment variable is also set to one of the openai models in the dictionaries. If the env variable is not set, then original function behaviors, of checking the dictionaries using the functions model_name variable, are the default.